added behavior img in front, fixed cleaning variables after fetchs, move multer to midleware
