+++ public/style.css
@@ -94,6 +94,7 @@ h3 {
     font-size: 19px;
     border-radius: 3%;
 }
+
 .my_coment_button {
     width: fit-content;
     padding: 6px;
@@ -109,7 +110,7 @@ main {
 
 .task {
     display: flex;
-    flex-direction: row;
+    flex-direction: row-reverse;
     flex-wrap: wrap;
     padding: 2px 5px;
     border: 2px solid rgba(0, 0, 0, .08);
@@ -123,7 +124,15 @@ main {
 
 .task_content {
     width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
 }
+
+.task_content a {
+    align-self: end;
+}
+
 .task_face {
     display: inline-block;
     line-height: 24px;
@@ -135,7 +144,7 @@ main {
     border: 3px solid white;
     border-radius: 50%;
     padding: 0.25rem;
-    font-family: Apple Color Emoji,Segoe UI Emoji,Noto Color Emoji,Android Emoji,EmojiSymbols,EmojiOne Mozilla,Twemoji Mozilla,Segoe UI Symbol,Noto Color Emoji Compat,emoji,var(--font-noto-emojipedia-fallback);
+    font-family: Apple Color Emoji, Segoe UI Emoji, Noto Color Emoji, Android Emoji, EmojiSymbols, EmojiOne Mozilla, Twemoji Mozilla, Segoe UI Symbol, Noto Color Emoji Compat, emoji, var(--font-noto-emojipedia-fallback);
     font-size: 25px;
 }
 
@@ -145,18 +154,19 @@ button {
     padding-left: 3px;
     padding-right: 3px;
 }
+
 .button.check {
-  display: flex;
+    display: flex;
 }
 
 .status h3 {
     padding: 16px;
 }
 
-.contentText {
+.headerContent, .corectPost {
     display: flex;
     flex-direction: row;
-    background-color: rgb(245,246,247);
+    background-color: rgb(245, 246, 247);
     width: 100%;
 }
 
@@ -168,7 +178,8 @@ button {
 
 .wrapper {
     background: #fff;
-    position: relative;
+    position: static;
+    /* position: relative;*/
     box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.2),
     0 25px 50px 0 rgba(0, 0, 0, 0.1);
     padding: 25px;
@@ -257,26 +268,30 @@ hr {
     gap: 17px;
     width: 80vw;
 }
-.wrapper>.wrapper.pagination {
+
+.wrapper > .wrapper.pagination {
     display: flex;
     justify-content: center;
     gap: 5px;
     margin-top: 30px;
     padding: 5px;
 }
-.LoginInput .errorMsg{
+
+.LoginInput .errorMsg {
     color: darkred;
     top: 116px;
     position: absolute;
 }
-.LoginInput .errorMsg.pass{
+
+.LoginInput .errorMsg.pass {
     top: 180px;
 }
-.LoginInput .errorMsg.user-name{
+
+.LoginInput .errorMsg.user-name {
     top: 245px;
 }
 
-.wrapper .task .commits-section{
+.wrapper .task .commits-section {
     margin-left: 50px;
     display: flex;
     flex-direction: column;
@@ -285,18 +300,115 @@ hr {
     flex-wrap: nowrap;
     border: 2px solid rgba(0, 0, 0, .08);
 }
-.wrapper .task .commits-section .contentText{
+
+.wrapper .task .commits-section .headerContent {
     padding: 0px 5px;
 }
-.wrapper div> label{
+
+.wrapper div > label {
     justify-content: space-between;
 }
-.wrapper div> label input:nth-child(1)  {
-    width: 80% !important;
+
+
+
+/*modal window for img*/
+.lightboxOverlay {
+    position: absolute;
+    top: 0;
+    left: 0;
+    z-index: 9999;
+    background-color: #000;
+    filter: alpha(Opacity=80);
+    opacity: .7;
+    visibility: hidden;
+    transition-property: all;
+    transition-duration: 0.5s;
+    transition-delay: 0s;
+    transition-timing-function: ease-in-out;
 }
-.wrapper div> label input:nth-child(2)  {
-    width: 15% !important;
+
+
+ .lb-image {
+    width: 320px;
+    height: 240px;
 }
-.wrapper div> label SPAN  {
-    width: 15%;
+
+.lb-data .lb-close {
+    display: block;
+    float: right;
+    width: 30px;
+    height: 30px;
+    background: url(../static/close.png) top right no-repeat;
+    text-align: right;
+    outline: 0;
+    filter: alpha(Opacity=70);
+    opacity: .7;
+    -webkit-transition: opacity .2s;
+    -moz-transition: opacity .2s;
+    -o-transition: opacity .2s;
+    transition: opacity .2s;
+    cursor: pointer;
 }
+
+.lb-data .lb-details {
+    width: 85%;
+    float: left;
+    text-align: left;
+    line-height: 1.1em;
+}
+.lb-data {
+    padding: 0 4px;
+    color: #ccc;
+}
+.lb-dataContainer {
+    margin: 0 auto;
+    padding-top: 5px;
+    width: 100%;
+    border-bottom-left-radius: 4px;
+    border-bottom-right-radius: 4px;
+    width: 320px;
+}
+.lb-loader {
+    top: 43%;
+    height: 25%;
+    width: 100%;
+}
+.lb-nav {
+    position: absolute;
+    top: 0;
+    left: 0;
+    height: 100%;
+    width: 100%;
+    z-index: 10;
+}
+.lightbox .lb-image {
+    display: block;
+    height: auto;
+    max-width: inherit;
+    max-height: none;
+    border-radius: 3px;
+    border: 4px solid #fff;
+}
+.lightbox {
+    width: 100%;
+    z-index: 10000;
+    font-weight: 400;
+    outline: 0;
+    position: absolute;
+    top: 300px;
+    left: 0;
+     visibility: hidden;
+    transition-property: all;
+    transition-duration: 0.5s;
+    transition-delay: 0s;
+    transition-timing-function: ease-in-out;
+}
+.lb-outerContainer {
+    position: relative;
+    margin: 0 auto;
+    border-radius: 4px;
+    background-color: #fff;
+    width: 320px;
+    height: 240px;
+}
+
