+++ src/midleware/iscorectToken.ts
@@ -6,7 +6,7 @@ export function isCorrectToken(req: Request, res: Response, next: NextFunction) 
     const token = _csrfToken ? _csrfToken : XSRF;
 
     const tokens:Tokens = new Tokens();
-   /* console.log('token.body-',req.body);
+    /*console.log('token.body-',req.body);
     console.log('req.session.secretForCustomer-',req.session.secretForCustomer);*/
 
     if (!req.session.secretForCustomer || !tokens.verify(req.session.secretForCustomer, token)) {
